native 16bit int and large floats fix
On a processor where an int is 16 bit, the max float value will be capped to 32767 due to type conversions. The PRINTF_MAX_FLOAT define is meant to prevent the limitation being exposed but it was written with 32 bit ints in mind. Explicitly setting the conversion type to int32_t fixes this. Added a note in readme against the define as a gotcha